Extract MessageLog banner builder to lib/messageLog
